Fix typos and rename routes
